Retrieve data from DB error page pending data and server list

+++ src/containers/ErrorsPage/ErrorsPage.js
@@ -1,56 +1,82 @@
 import React, { Component } from "react";
+
+import axios from '../../services/axios-orders';
+
 import Auxiliary from "../../hoc/Auxiliary";
 import ErrorControls from "./ErrorControls/ErrorControls";
 import ErrorsList from "./ErrorsList/ErrorsList";
+import Spiner from '../../components/UI/Spinner/Spinner';
 
 class ErrorsPage extends Component {
   state = {
-    errorList: 
-      [
-            {code: 1, name:"name1", description: "Error Description number 1 ...", type:"1", dateTime:'1/1/2020'},
-            {code: 2, name:"name2", description: "fdfdfd fdfdf fdfd", type:"2", dateTime:'1/1/2020'},
-            {code: 3, name:"name3", description: "fdfdfd fdfdf fdfd", type:"3", dateTime:'1/1/2020'},
-            {code: 4, name:"name4", description: "fdfdfd fdfdf fdfd", type:"4", dateTime:'1/1/2020'},
-            {code: 5, name:"name5", description: "fdfdfd fdfdf fdfd", type:"5", dateTime:'1/1/2020'},
-            {code: 6, name:"name6", description: "fdfdfd fdfdf fdfd", type:"5", dateTime:'1/1/2020'},
-            {code: 7, name:"name7", description: "fdfdfd fdfdf fdfd", type:"noerror", dateTime:'1/1/2020'},
-            {code: 8, name:"name8", description: "fdfdfd fdfdf fdfd", type:"noerror", dateTime:'1/1/2020'},
-            {code: 9, name:"name9", description: "fdfdfd fdfdf fdfd", type:"noerror", dateTime:'1/1/2020'}
-    ],
-    dumyCpunter: 0
+    errorList:null,
+    errorCounter: 0,
+    loading: false,
+    error: false
   };
+
+  componentDidMount() {
+    axios.get('errors.json').then(response => {
+      let currentErrorList = [];
+      for (let element in response.data) {
+        currentErrorList.push({
+          errorId: element,
+          errorCode: response.data[element].code,
+          errorName: response.data[element].name,
+          errorDescription: response.data[element].description,
+          errorType: response.data[element].type,
+          errorDateTime: response.data[element].datetime
+        });
+      }
+      this.setState({ errorCounter: currentErrorList.length });
+      this.setState({errorList:currentErrorList});
+    }).catch(error => {
+      console.log(error);
+      this.setState({ error: true });
+    });
+  }
 
   deleteAllHandler = () => {
     console.log("deleteAllHandler function");
     const oldErrorList = this.state.errorList;
     console.log(oldErrorList);
     const newErrorList = [];
-    this.setState({errorList: newErrorList});
+    this.setState({ errorList: newErrorList });
+    // Delete all from errors in database and receive 200 OK change state
   }
 
   deleteErrorHandler = (deleteId) => {
     console.log("deleteErrorHandler function " + deleteId);
     const currentErrorList = [...this.state.errorList]; // για να το κάνω με emutable τρόπο
-    currentErrorList.splice(deleteId,1);
-    this.setState({errorList : currentErrorList});
+    currentErrorList.splice(deleteId, 1);
+    this.setState({ errorList: currentErrorList });
+    // Delete one from DB with ID and receive 200 ok change state
   }
 
-  editErrorHandler = (editId) => {
-    console.log("editErrorHandler function " + editId);
-  }
 
   render() {
-    return (
-      <Auxiliary>
-        <ErrorsList 
+    let errorPage = null;
+    if (this.state.errorList) {
+      errorPage = <span>
+        <ErrorsList
           errors={this.state.errorList}
           deleteError={(index) => this.deleteErrorHandler(index)}
-          editError={(index) => this.editErrorHandler(index)}
         />
-        <ErrorControls 
-          numOfErrors={this.state.errorList.length} 
+        <ErrorControls
+          numOfErrors={this.state.errorList.length}
           deleteAll={this.deleteAllHandler}
         />
+      </span>
+    }else if(this.state.error){
+      errorPage = <p>Errors can't de loaded!</p>
+    }else {
+      errorPage= <Spiner/>
+    }
+      
+
+    return (
+      <Auxiliary>
+        {errorPage}
       </Auxiliary>
     );
   }
